testproject/requirements.txt
Django==4.2.5
psycopg2-binary>=2.9,<3.0
djangorestframework==3.14.0
mysqlclient==2.2.0